# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      success:
        type: boolean
    type: object
...
  models.Notification:
    properties:
      created_at:
        type: string
      data:
        additionalProperties: true
        type: object
      id:
        type: string
      is_read:
        type: boolean
      is_sent:
        type: boolean
      message:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/models.NotificationType'
      user_id:
        type: string
    type: object
  models.NotificationPreferences:
    properties:
      created_at:
        type: string
      email_enabled:
        type: boolean
      id:
        type: string
      marketing_emails:
        type: boolean
      match_notifications:
        type: boolean
      message_notifications:
        type: boolean
      profile_notifications:
        type: boolean
      push_enabled:
        type: boolean
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.NotificationType:
    enum:
    - new_match
    - super_like
    - match_expiring
    - new_message
    - message_read
    - profile_viewed
    - photo_liked
    - account_verified
    - account_warning
    - promotional
    type: string
    x-enum-varnames:
    - NotificationNewMatch
    - NotificationSuperLike
    - NotificationMatchExpiring
    - NotificationNewMessage
    - NotificationMessageRead
    - NotificationProfileViewed
    - NotificationPhotoLiked
    - NotificationAccountVerified
    - NotificationAccountWarning
    - NotificationPromotional
...
  models.UpdateNotificationPreferencesRequest:
    properties:
      email_enabled:
        type: boolean
      marketing_emails:
        type: boolean
      match_notifications:
        type: boolean
      message_notifications:
        type: boolean
      profile_notifications:
        type: boolean
      push_enabled:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@matching-api.example.com
    name: API Support
    url: https://matching-api.example.com/support
  description: A complete Go-based REST API for a dating application with real-time
    chat, sophisticated matching algorithms, and comprehensive user management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://matching-api.example.com/terms
  title: Matching API - Matching App Backend
  version: "1.0"
paths:
...
  /notifications:
    get:
      consumes:
      - application/json
      description: Retrieve user's notifications with pagination and filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of notifications per page (max 50)
        in: query
        name: limit
        type: integer
      - description: Filter by notification type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notifications retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  properties:
                    limit:
                      type: integer
                    notifications:
                      items:
                        $ref: '#/definitions/models.Notification'
                      type: array
                    page:
                      type: integer
                    total:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad request - invalid pagination
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user notifications
      tags:
      - Notifications
  /notifications/{notificationID}/read:
    put:
      consumes:
      - application/json
      description: Mark a specific notification as read for the current user
      parameters:
      - description: Notification ID
        in: path
        name: notificationID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notification marked as read
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad request - missing notification ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Notification not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - Notifications
...
  /notifications/preferences:
    get:
      consumes:
      - application/json
      description: Retrieve the current user's notification preferences
      produces:
      - application/json
      responses:
        "200":
          description: Notification preferences retrieved
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  properties:
                    preferences:
                      $ref: '#/definitions/models.NotificationPreferences'
                  type: object
              type: object
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get notification preferences
      tags:
      - Notifications
    put:
      consumes:
      - application/json
      description: Update the current user's notification preferences
      parameters:
      - description: Notification preferences update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateNotificationPreferencesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Notification preferences updated
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Notification preferences not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update notification preferences
      tags:
      - Notifications
  /notifications/read-all:
    put:
      consumes:
      - application/json
      description: Mark all notifications as read for the current user
      produces:
      - application/json
      responses:
        "200":
          description: All notifications marked as read
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  properties:
                    notifications_updated:
                      type: integer
                  type: object
              type: object
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - Notifications
...
  /notifications/unread-count:
    get:
      consumes:
      - application/json
      description: Get the count of unread notifications for the current user
      produces:
      - application/json
      responses:
        "200":
          description: Unread count retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  properties:
                    unread_count:
                      type: integer
                  type: object
              type: object
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get unread notification count
      tags:
      - Notifications
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"